Fixing compilation for VS

+++ trunk/src/PotatoEngine/Core/Math.cpp
@@ -1,3 +1,4 @@
+#include "../stdafx.h"
 #include "Math.hpp"
 #include "LibsHelpers.hpp"
 
@@ -9,12 +10,12 @@ const double Math::DEG_2_RAD = Math::precisePi() / 180.;
 
 float Math::rad2deg(float angle)
 {
-	return angle * RAD_2_DEG;
+	return angle * static_cast<float>(RAD_2_DEG);
 }
 
 float Math::deg2rad(float angle)
 {
-	return angle * DEG_2_RAD;
+	return angle * static_cast<float>(DEG_2_RAD);
 }
 
 double Math::precisePi()
@@ -28,19 +29,19 @@ double Math::precisePi()
 
 float Math::pi()
 {
-	return (float)precisePi();
+	return static_cast<float>(precisePi());
 }
 
 float Math::lerp(float source, float target, float coef)
 {
-	coef = clamp<float>(0., 1., coef);
-	return (1. - coef) * source + coef * target;
+	coef = clamp<float>(0.f, 1.f, coef);
+	return (1.f - coef) * source + coef * target;
 }
 
 Vector2f Math::lerp(const Vector2f& source, const Vector2f& target, float coef)
 {
-	coef = clamp<float>(0., 1., coef);
-	return (1. - coef) * source + coef * target;
+	coef = clamp<float>(0.f, 1.f, coef);
+	return (1.f - coef) * source + coef * target;
 }
 
 }
